Add integration test using trellis-cli
Uses trellis-cli to create a new Trellis project from the current SHA.
The project is provisioned and deployed all using trellis-cli.

Finally the site is verified that WordPress was actually installed and
working.

The steps are a little more complex than someone provisioning a normal
remote server for a few reasons:

1. GitHub's Ubuntu image has a lot of software installed by default
   including mysql-server which needs to be purged completely.
2. there's a few workarounds to force Ansible to use a local connection
   since we aren't actually SSHing into a remote server.